bench: 1 fix nps in bench

+++ engine/main.cpp
@@ -14,7 +14,7 @@ int main(int argc, char *argv[]) {
 		uint64_t start = clock();
 		search(board, 1000, 1);
 		uint64_t end = clock();
-		std::cout << 1 << " nodes " << (ngames() / ((double)(end - start) / CLOCKS_PER_SEC)) << " nps" << std::endl;
+		std::cout << 1 << " nodes " << (int)(ngames() / ((double)(end - start) / CLOCKS_PER_SEC)) << " nps" << std::endl;
 		return 0;
 	}
 	bool online = argc == 2 && std::string(argv[1]) == "--online";
